Various updates to format conversions.

+++ FluxSolution/BaseLibrary/Source/Types/Stopwatch/GetMicroseconds.em.cs
@@ -4,6 +4,6 @@ namespace Flux
 	{
 		/// <summary>Compute how many microseconds has elapsed.</summary>
 		public static double GetTotalMicroseconds(this System.Diagnostics.Stopwatch source)
-			=> (double)(source ?? throw new System.ArgumentNullException(nameof(source))).ElapsedTicks / (double)System.Diagnostics.Stopwatch.Frequency * 1000000.0;
+			=> (source ?? throw new System.ArgumentNullException(nameof(source))).ElapsedTicks / (double)System.Diagnostics.Stopwatch.Frequency * 1000000.0;
 	}
 }
