+++ FluxSolution/BaseLibrary/Source/Types/Stopwatch/GetNanoseconds.em.cs
@@ -4,6 +4,6 @@ namespace Flux
 	{
 		/// <summary>Compute how many nanoseconds has elapsed.</summary>
 		public static double GetTotalNanoseconds(this System.Diagnostics.Stopwatch source)
-			=> (double)(source ?? throw new System.ArgumentNullException(nameof(source))).ElapsedTicks / (double)System.Diagnostics.Stopwatch.Frequency * 1000000000.0;
+			=> (source ?? throw new System.ArgumentNullException(nameof(source))).ElapsedTicks / (double)System.Diagnostics.Stopwatch.Frequency * 1000000000.0;
 	}
 }
